+++ src/pages/LoginPage/components/LogoSection.tsx
@@ -1,0 +1,15 @@
+const LogoSection = () => {
+  return (
+    <section>
+      <h1 className="w-52 h-52 rounded-full overflow-hidden m-auto">
+        <img
+          src="https://placehold.co/200x200?text=haertz"
+          alt="logo image"
+          className="size-full object-cover"
+        />
+      </h1>
+    </section>
+  );
+};
+
+export default LogoSection;
